third_party/WebKit/LayoutTests/fast/dom/document-set-title-no-child-on-empty.html

<!DOCTYPE html>
<title>document title and the empty string</title>
<script src="../../resources/js-test.js"></script>
<body>
<script>
    description("On setting document.title to the empty string, no text node must be created");

    var head = document.documentElement.firstChild;
    head.removeChild(head.firstChild);
    shouldBeEqualToString("document.title", "");
    document.title = "";
    shouldBeEqualToString("document.title", "");
    shouldBeTrue("head.lastChild instanceof HTMLTitleElement");
    shouldBe("head.lastChild.firstChild", 'null');
</script>
</body>
