third_party/blink/web_tests/fast/dom/document-set-title-no-child-on-empty-expected.txt

On setting document.title to the empty string, no text node must be created

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.title is ""
PASS document.title is ""
PASS head.lastChild instanceof HTMLTitleElement is true
PASS head.lastChild.firstChild is null
PASS successfullyParsed is true

TEST COMPLETE
